+++ tasks/ec2_ssh.js
@@ -19,37 +19,7 @@ module.exports = function(grunt){
         var done = this.async();
 
         ssh.connect({ name: name }, function (c) {
-
-            grunt.log.ok('Connection established! Use ctrl+c twice to exit ssh session');
-
-            var cancelled = false;
-            var interactive = ssh.interactive(c);
-
-            process.stdin.resume();
-            process.stdin.on('data', function (data) {
-                cancelled = false;
-                interactive.write(data);
-            });
-
-            process.on('SIGINT', function() {
-                if (!cancelled) {
-
-                    interactive.cancel(function () {
-                        grunt.log.write('\nEnter %s again to exit session', chalk.red('ctrl+c'));
-                        cancelled = true;
-                    });
-                }
-
-                process.once('SIGINT', function() {
-                    if (cancelled) {
-                        cancelled = false;
-                        grunt.log.writeln();
-                        grunt.log.ok('Exiting ssh session...');
-                        c.end();
-                    }
-                });
-            });
-
+            ssh.interactive(c);
         }, done);
 
     });